+++ Changelog.md
@@ -145,4 +145,8 @@
 |27.10.2023|Get-AllVoiceAppsAndResourceAccounts.ps1|1.0.1|Fix result caching by making all relevant variables global.|
 |27.10.2023|M365CallFlowVisualizerV2.ps1|3.0.11|Add support for PDF Export. Thanks @MicheleBomello for this PR!|
 |06.11.2023|M365CallFlowVisualizerV2.ps1|3.1.0|Add support to check if an Auto Attendant is currently in business- or after hours schedule or in holiday schedule by using -CheckCallFlowRouting.|
-|06.11.2023|M365CallFlowVisualizerV2.ps1|3.1.1|Bug fixes, improvements in console and diagram output (support for ComplementEnabled: False schedules)|
+|07.11.2023|M365CallFlowVisualizerV2.ps1|3.1.1|Bug fixes, improvements in console and diagram output (support for ComplementEnabled: False schedules)|
+|07.11.2023|M365CallFlowVisualizerV2.ps1|3.1.2|Remove (broken) support for ServicePrincipal auth. Add warning message when -ConnectWithServicePrincipal is used.|
+|07.11.2023|Get-M365CFVTeamsAdminToken.ps1|1.0.0|File removed.|
+|07.11.2023|Connect-M365CFV.ps1|1.1.1|Remove support for Teams and Graph sign in via Entra ID App Registration / Service Principal. This has been moved into its own function.|
+|07.11.2023|Connect-MsTeamsServicePrincipal.ps1|1.0.0|Prepare for support for Teams and Graph sign in via Entra ID App Registration / Service Principal once it supports Get-CsOnlineApplicationInstance.|